requirements.txt
streamlit==1.31.0
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
rdkit-pypi==2022.9.5
matplotlib==3.7.2
seaborn==0.12.2
Pillow==10.0.0
xgboost==2.0.3
